requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-community
langgraph
faiss-cpu
openai
pillow
python-dotenv
requests

